One runner for all layers

+++ retroarch/build-arch.sh
@@ -61,44 +61,6 @@ function fetch_retroarch()
   rm -rf squashfs-root
 }
 
-# Fetch & configure flatimage
-function fetch_flatimage()
-{
-  msg "${IMAGE:?IMAGE is undefined}"
-
-  if [[ -n "$1" ]]; then
-    cp "$1" "$IMAGE"
-  else
-    wget "$(wget -qO - "https://api.github.com/repos/ruanformigoni/flatimage/releases/latest" \
-      | jq -r '.assets.[].browser_download_url | match(".*arch.flatimage$").string')"
-    chmod +x "$IMAGE"
-  fi
-
-  # Enable network
-  "$IMAGE" fim-perms set network
-
-  # Update
-  "$IMAGE" fim-root pacman -Syu --noconfirm
-
-  # Install dependencies
-  "$IMAGE" fim-root pacman -S libxkbcommon libxkbcommon-x11 \
-    lib32-libxkbcommon lib32-libxkbcommon-x11 libsm lib32-libsm fontconfig \
-    libxinerama lib32-libxinerama \
-    lib32-fontconfig noto-fonts sdl2 lib32-sdl2 --noconfirm
-
-  # Install video packages
-  "$IMAGE" fim-root pacman -S xorg-server mesa lib32-mesa \
-    glxinfo pcre xf86-video-amdgpu vulkan-radeon lib32-vulkan-radeon \
-    xf86-video-intel vulkan-intel lib32-vulkan-intel vulkan-tools --noconfirm
-
-  # Gameimage dependencies
-  "$IMAGE" fim-root pacman -S noto-fonts libappindicator-gtk3 \
-    lib32-libappindicator-gtk3 --noconfirm
-
-  # Commit changes
-  "$IMAGE" fim-commit
-}
-
 # Create layer filesystems for retroarch and its assets
 function compress_retroarch()
 {
@@ -123,65 +85,32 @@ function compress_retroarch()
 
 function main()
 {
+  export IMAGE="$1"
+  if ! [ -f "$IMAGE" ]; then
+    echo "Please specify a regular file as the image path"
+    exit 1
+  fi
+
   # shellcheck disable=2155
   export SCRIPT_DIR="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd)"
   export BUILD_DIR="$SCRIPT_DIR/build"
 
   rm -rf "$BUILD_DIR"; mkdir "$BUILD_DIR"; cd "$BUILD_DIR"
 
-  # Container file path
-  IMAGE="$BUILD_DIR/arch.flatimage"
-
   # Retroarch
   fetch_retroarch
-
-  # FlatImage
-  if [[ "$1" = "--flatimage" ]] && [[ -n "$2" ]]; then
-    fetch_flatimage "$2"
-  else
-    fetch_flatimage
-  fi
 
   # Include retroarch in flatimage
   compress_retroarch
 
-  # Create directories
-  "$IMAGE" fim-exec sh -c 'mkdir -p /home/gameimage/.config'
-  "$IMAGE" fim-exec sh -c 'mkdir -p /home/gameimage/.local/share'
-
-  # Set environment variables
-  "$IMAGE" fim-env set 'PATH="/opt/retroarch/data/bin:$PATH"' \
-    'FIM_BINARY_RETROARCH="/opt/retroarch/boot"' \
-    'HOME=/home/gameimage' \
-    'XDG_CONFIG_HOME=/home/gameimage/.config' \
-    'XDG_DATA_HOME=/home/gameimage/.local/share'
-
-  # Set default command
-  "$IMAGE" fim-boot '/opt/retroarch/boot'
-
-  # Set perms
-  "$IMAGE" fim-perms set home,media,audio,wayland,xorg,dbus_user,dbus_system,udev,usb,input,gpu,network
-
-  # Save changes
-  "$IMAGE" fim-commit
-
-  # Rename
-  mv "$IMAGE" retroarch.flatimage
-
   # Create dist
-  mkdir -p ../dist
+  mkdir -p ../dist && cd ../dist
 
   # Move to-be-released files
-  mv retroarch.flatimage ../dist
-  mv retroarch.layer ../dist
-
-  # Enter dist dir
-  cd ../dist
+  mv "$BUILD_DIR"/retroarch.layer .
 
   # Create checksum for everything
-  for i in *; do
-    sha256sum "$i" > "${i}.sha256sum"
-  done
+  sha256sum retroarch.layer > retroarch.layer.sha256sum
 }
 
 main "$@"
